refactor(be): move check_session_authorization into authz_utils
`check_session_authorization`, `CallerCapability` and the `Sealed`
seal token belong with the other caller-against-anchor authorization
primitives in `authz_utils.rs`, alongside `check_authorization`. This
mirrors the existing pattern: email-recovery's principal derivation
also lives in `authz_utils.rs::check_authorization`, pulling the seed
function from `email_recovery::smtp`.

`session_delegation.rs` keeps the session-specific primitives
(`session_delegation_seed`, `expected_session_principal`) as
`pub(crate)` so the authz module can call into them, plus the two
endpoint helpers (`prepare_session_delegation`,
`get_session_delegation`).

No behavior change. The seal property is preserved -- `Sealed`'s
private field still makes `check_session_authorization` the only
mint site for a `CallerCapability`.
