+++ src/internet_identity/src/authz_utils.rs
@@ -1,5 +1,6 @@
 use crate::anchor_management::post_operation_bookkeeping;
 use crate::ii_domain::IIDomain;
+use crate::session_delegation::expected_session_principal;
 use crate::storage::anchor::Anchor;
 use crate::storage::StorageError;
 use crate::{anchor_management, state};
@@ -10,6 +11,17 @@ use internet_identity_interface::internet_identity::types::{
     AnchorNumber, AuthorizationKey, IdentityNumber,
 };
 use std::fmt::{Display, Formatter};
+
+/// Outcome of [`check_session_authorization`]. The wrapped `Sealed` token has a
+/// private field, so a `CallerCapability` can only be built inside this module —
+/// making this function the only way to obtain scoped authorization.
+pub enum CallerCapability {
+    FullAuth(Box<Anchor>, AuthorizationKey, Sealed),
+    SessionScoped(Sealed),
+}
+
+pub struct Sealed(#[allow(dead_code)] ());
+
 
 #[derive(Debug)]
 pub enum IdentityUpdateError {
@@ -190,6 +202,28 @@ pub fn is_self_authenticating(principal: Principal) -> bool {
     let self_authenticating_bytes = self_authenticating.as_slice();
 
     principal_bytes.last() == self_authenticating_bytes.last()
+}
+
+/// Like [`check_authorization`], but also accepts the session-scoped
+/// principal derived for `anchor_number` (see [`crate::session_delegation`]).
+/// Endpoints opting into session callers go through this gate instead.
+pub fn check_session_authorization(
+    anchor_number: AnchorNumber,
+) -> Result<CallerCapability, AuthorizationError> {
+    if let Ok((anchor, key)) = check_authorization(anchor_number) {
+        return Ok(CallerCapability::FullAuth(
+            Box::new(anchor),
+            key,
+            Sealed(()),
+        ));
+    }
+
+    let salt_initialised = state::storage_borrow(|storage| storage.salt().is_some());
+    if salt_initialised && caller() == expected_session_principal(anchor_number) {
+        return Ok(CallerCapability::SessionScoped(Sealed(())));
+    }
+
+    Err(AuthorizationError::from(caller()))
 }
 
 #[cfg(test)]
